+++ src/python/arcor2_fit_demo/object_types/abstract_dobot.py
@@ -1,4 +1,3 @@
-import copy
 from dataclasses import dataclass
 from typing import List, Optional, Set, cast
 
@@ -131,9 +130,6 @@ class AbstractDobot(Robot):
         :return:
         """
 
-        # TODO this is a temporal workaround until the generated ActionPoints class will return copies of poses
-        pick_pose = copy.deepcopy(pick_pose)
-
         pick_pose.position.z += vertical_offset
         self.move(pick_pose, MoveType.JOINTS)  # pre-pick pose
         pick_pose.position.z -= vertical_offset
@@ -149,9 +145,6 @@ class AbstractDobot(Robot):
         :param vertical_offset:
         :return:
         """
-
-        # TODO this is a temporal workaround until the generated ActionPoints class will return copies of poses
-        place_pose = copy.deepcopy(place_pose)
 
         place_pose.position.z += vertical_offset
         self.move(place_pose, MoveType.JOINTS)  # pre-place pose